Skip tests with failPoint when the server does not provide this feature
